% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCompanion.R
\name{betaParams}
\alias{betaParams}
\title{Compute parameters of beta distribution}
\usage{
betaParams(
  a = NA,
  b = NA,
  mean = NA,
  sd = NA,
  k = NA,
  mode = NA,
  verbose = FALSE
)
}
\arguments{
\item{a}{shape parameter}

\item{b}{shape parameter}

\item{mean}{mean of the distribution}

\item{sd}{standard deviation of the distribution}

\item{k}{concentration of the distribution}

\item{mode}{of the distribution}

\item{verbose}{= FALSE set to TRUE to get more explanation}
}
\value{
A list of all possible values of parameters
}
\description{
Compute $a$ and $b$ params of beta distribution
from other meaningfull params.
}
\examples{
betaParams(mean=0.5, k=8)
betaParams(mode=0.6, k=8)
betaParams(mean=0.5, s=0.2)
betaParams(a=12, b=12)

}
\author{
Julien Bousquet (2022)
}
