% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCompanion.R
\name{betaShiny}
\alias{betaShiny}
\title{Demonstrate behaviour of the parameters of
beta distribution, in a shiny application.}
\usage{
betaShiny()
}
\value{
None
}
\description{
Demonstrate behaviour of the parameters of
beta distribution, in a shiny application.
}
\examples{
betaShiny()

}
\author{
Julien Bousquet (2022)
}
