% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{crosscorr}
\alias{crosscorr}
\title{Add legend on crosscor.plot()}
\usage{
crosscorr(MCMC)
}
\arguments{
\item{MCMC}{an MCMC sample from JAGS}
}
\description{
This function use crosscor.plot() from coda, and add the
variables' names above the diagonal. Help understand
correlations problems in MCMC.
}
\examples{
#  no example
}
\author{
Julien Bousquet (2022)
}
