% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{DbdaDensPlot}
\alias{DbdaDensPlot}
\title{Density Plot.}
\usage{
DbdaDensPlot(codaObject, parName = varnames(codaObject)[1], plColors = NULL)
}
\arguments{
\item{codaObject, }{coda object from an MCMC sampler.}

\item{parName, }{string giving the name of the parameter to plot.}

\item{plColors, }{vector of strings giving the colors of the different chains plot.}
}
\description{
Function for plotting Density graphics, usually for MCMC. Used in diagMCMC().
Writen by John K. Kruschke in Doing Bayesian Data Analysis, 2nd edition, 2015
}
\author{
John K. Kruschke (2015)
}
