% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{diagMCMC}
\alias{diagMCMC}
\title{Graphics for Diagnostic of MCMC chains' convergence, in coda format.}
\usage{
diagMCMC(
  codaObject,
  parName = varnames(codaObject)[1],
  saveName = NULL,
  saveType = "jpg"
)
}
\arguments{
\item{codaObject, }{coda object from an MCMC sampler.}

\item{parName, }{string giving the name of the parameter to plot.}

\item{saveName, }{a string giving the name of the file to write the image.}

\item{saveType, }{a string giving the saving format of the image. : 'jpg', 'png', 'eps'.}
}
\description{
Usual diagnostic statistics to assess convergence of MCMC chains : traceplot, Gelman shrinkage
coefficient, autocorrelation, and density plots.
Writen by John K. Kruschke in Doing Bayesian Data Analysis, 2nd edition, 2015
}
\author{
John K. Kruschke (2015)
}
