% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCompanion.R
\name{gammaShiny}
\alias{gammaShiny}
\title{Demonstrate behaviour of the parameters of
Weibull distribution, in a shiny application.}
\usage{
gammaShiny()
}
\value{
None
}
\description{
Demonstrate behaviour of the parameters of
Weibull distribution, in a shiny application.
}
\examples{
cat('gammaShiny()')

}
\author{
Julien Bousquet (2022)
}
