% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCompanion.R
\name{normaleShiny}
\alias{normaleShiny}
\title{Shiny application for normal distribution}
\usage{
normaleShiny()
}
\value{
None
}
\description{
Demonstrate behaviour of the parameters of
normal distribution, in a shiny application.
}
\examples{
normaleShiny()

}
\author{
Julien Bousquet (2022)
}
