% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{openGraph}
\alias{openGraph}
\title{Function for opening a graphic device.}
\usage{
openGraph(width = 7, height = 7, mag = 1, ...)
}
\arguments{
\item{width}{width of the device in inches.}

\item{height}{height of the device in inches.}

\item{mag}{magnification coefficient.}

\item{...}{more parameters for \code{dev.new()}}
}
\description{
Function for opening graphics that operate the same for
Windows and Macintosh and Linux operating systems.
At least, that's the hope! writen by
John K. Kruschke in Doing Bayesian Data Analysis, 2nd edition, 2015
}
\examples{
#openGraph()
hist(rnorm(2000))
}
\author{
John K. Kruschke (2015)
}
