% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{saveGraph}
\alias{saveGraph}
\title{Function for saving a graphic device.}
\usage{
saveGraph(file = "saveGraphOutput", type = "pdf", ...)
}
\arguments{
\item{file, }{string of the name of file.}

\item{type, }{string of the file type. 'pdf' by default, 'jpg', 'png' or 'eps' are possible.}

\item{..., }{more parameters for \code{savePlot()}}
}
\description{
Function for saving graphics that operate the same for
Windows and Macintosh and Linux operating systems.
Writen by John K. Kruschke in Doing Bayesian Data Analysis, 2nd edition, 2015
}
\examples{
#openGraph()
hist(rnorm(2000))
#saveGraph(file='myfile', type='pdf')
}
\author{
John K. Kruschke (2015)
}
