% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCompanion.R
\name{StudentShiny}
\alias{StudentShiny}
\title{Demonstrate behaviour of the parameters of
Student distribution, in a shiny application.}
\usage{
StudentShiny()
}
\value{
None
}
\description{
Demonstrate behaviour of the parameters of
Student distribution, in a shiny application.
}
\examples{
StudentShiny()

}
\author{
Julien Bousquet (2022)
}
