% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{summarizePost}
\alias{summarizePost}
\title{Summarize posterior MCMC}
\usage{
summarizePost(paramSampleVec, compVal = NULL, ROPE = NULL, credMass = 0.95)
}
\arguments{
\item{paramSampleVec}{the vector of the MCMC parameter's sample.}

\item{compVal}{numeric : value to compare.}

\item{ROPE}{: vector of 2 limits defining the Range Of Practical Equivalence.}

\item{credMass, }{the mass of credibility in the High Density Interval, 0.95 by default.}
}
\value{
the vector of mcmc summary values.
}
\description{
Summarizing distribution of a large sample;
typically applied to MCMC posterior, writen by
John K. Kruschke in Doing Bayesian Data Analysis, 2nd edition
}
\examples{
summarizePost(rnorm(2000), ROPE=c(-1,1), compVal=0)
}
\author{
John K. Kruschke (2015)
}
