% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCompanion.R
\name{WeibullShiny}
\alias{WeibullShiny}
\title{Demonstrate behaviour of the parameters of
Weibull distribution, in a shiny application.}
\usage{
WeibullShiny()
}
\value{
None
}
\description{
Demonstrate behaviour of the parameters of
Weibull distribution, in a shiny application.
}
\examples{
cat('WeibullShiny()')

}
\author{
Julien Bousquet (2022)
}
